计算热流，初步确认无误
